Move getDatastore util to tests.
